// Types/UtilTypes.ts
export enum UtilTypes {
    ZOOM_IN = "ZOOM_IN",
    ZOOM_OUT = "ZOOM_OUT",
    MEASUREMENT = "MEASUREMENT",
    HIDE = "HIDE",
    SHARPEN = "SHARPEN",
    BLUR = "BLUR",
    GRAY_SCALE = "GRAY_SCALE",
    BRIGHTNESS = "BRIGHTNESS",
    CONTRAST = "CONTRAST",
    UNDO = "UNDO",
    REDO = "REDO"

}